Backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cloudinary==1.30.0
Django==4.1.2
djangorestframework==3.14.0
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
